split container test file
