+++ tests/playwright/sanity/modules/container/container-1.test.ts-snapshots/container-column-boxed-start-linux.png
