+++ tests/playwright/sanity/modules/container/container-1.test.ts-snapshots/container-column-full-start-linux.png
