+++ tests/playwright/sanity/modules/container/container-1.test.ts-snapshots/container-ltr-centered-linux.jpeg
