+++ tests/playwright/sanity/modules/container/container-2.test.ts-snapshots/container-dimensions-rtl-with-ltr-ui-linux.png
