+++ tests/playwright/sanity/modules/container/container-2.test.ts-snapshots/container-justify-controls-ltr-right-mobile-linux.jpeg
