+++ tests/playwright/sanity/modules/container/container-2.test.ts-snapshots/container-justify-controls-rtl-right-tablet-linux.jpeg
